feat(FTL-8789): update folder and variable names to generic one
